nbd: Allow exporting a volume without the backing chain
Add backing_chain argument to NBD server configuration. If false and
using volume in qcow2 format, expose only the specified volume instead
of the entire chain.

If not specified or true, export the entire chain. This argument has no
effect when volume format is raw.

Allocated areas in the volume will be read as data (zero=False) or as
zero (zero=True, hole=False). Unallocated areas will read as a hole
(zero=True, hole=True).

Here is an example for extents response:

    {"start": 0, "length": 65536, "zero": False, "hole": False},
    {"start": 65536, "length": 65536, "zero": True, "hole": False},
    {"start": 131072, "length": 65536, "zero": True, "hole": True},

- The first extent is data extent.
- The second extent is zero extents but it is not a hole. When copying
  this extent, the destination image must be zeroed.
- The last extent is a hole; when copying this extent, the destination
  image must not be zeroed.

Change-Id: I617c18a1a0639106f867f059d8baefca1e82f585
Bug-Url: https://bugzilla.redhat.com/1847090
Signed-off-by: Nir Soffer <nsoffer@redhat.com>