tests: Fix untrusted image verification tests
Older versions of qemu-img did not access the backing file when creating
a new image if image size was specified. Current version (5.1.0) tries to
access the backing file and fails with:

    Could not open '/path/to/file': No such file or directory
    Could not open backing image.

Change the tests to create the backing file, and use "qemu-img rebase"
to change the backing file instead of "qemu-img create".

Change-Id: I4e0c0bbc3b08a8017897696eb4b019cdaa7ffa1d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>