tool: sanlock: Prepare for configuring sanlock.conf
Clean up the messy isconfigured() by extracting small and simple
functions for checking if sanlock groups are configured, and if sanlock
needs to be restarted to apply groups changes.

Refactor isconfigured() and configure() to make it easy to check and
configure sanlock configuration file.

Add common logging when checking and configuring sanlock, matching other
configurators.

Change-Id: If87bc40707236e72a8362684e976e463378636e1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>