virt: check for virt-* commands during runtime
It is not a best idea to check for existence of the virt-* commands
during VDSM startup. When the commands are missing it is certainly bad,
as it means some of the VDSM dependencies were not properly installed,
but it should not prevent VDSM from starting up. Hence we drop the check
and let the invocations fail only when the functionality is actually
needed.

While at it drop the use of CommandPath as it does not have any value
when we look only for one fixed path.

Change-Id: Ie450a47629e7a638756ba314f152a461b1027d3b
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>